service_manager/pyproject.toml
[tool.poetry]
name = "service-manager-api"
version = "0.1.0"
description = ""
authors = ["Erika Fernanda <erika.fernanda@smartsuite.imd.ufrn.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
django-filter = "^23.3"
reportlab = "^4.0.6"
pyjwt = "^2.8.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
